Add step model and footprint for Murata inductor.
The step model doesn't include connections between pins and the inductor
coil in the middle, nor does it have the contour of the resin on the
bottom of the part. This is probably good enough, but I will need to get
a physical part if I want to add more detail (they're in the mail).