Move via off pad of C2 in AP5726 design.
Apparently, it's not good to put vias directly on pads (metal dome
switch trace design being an exception) because it may wick away solder
and also cause the surface to be uneven. An exception to this is when
you have a large thermal pad underneath an IC, in which case you may
want many vias spread out to provide surface area for solder to contact
as well as provide better heat distribution to the copper area on the
other side.