Remove NUM_REG modulo because it's not what we want.
It's not helpful to try and be clever when dealing with writing past the
intended register address. Just let it behave exactly as you would
expect.

Also I'm seeing a terrible bug in EEPROM. When the Attiny is programmed,
it will erase certain cells of the EEPROM (setting them back to 0xFF)
for some reason. This happens even after making sure "EEPROM retained"
is set, activating BOD, and burning those settings to fuses by doing
"Burn bootloader".

The good news is that everything works fine if you program, write the
EEPROM, and then read it. And it holds steady across resets. So I guess
it's just a programming issue.