Implement periodic eeprom writeback.
Simulate "watchdog" timer so that you only write back to eeprom after a
short time of no register changes (10 sec here). This will prolong the
lifetime of the eeprom cells by a lot, and also make it more responsive,
since it's not writing to eeprom all the time.

Include a second watchdog timer that writes back to eeprom even if there
is sustained register write activity. This timeout is 1 min. (So if
there has been any register activity, it will update the EEPROM after 1
min no matter what).