Remove lcd-peripherals and finish second draft of dsi signals.
The LCD peripherals doesn't seem like it adds enough to warrant having a
separate iteration for this pcb. The circuitry from the ap5726 and attiny
projects can hook up to the header pin on the dsi-signals board
directly.

This is a fully routed draft of the dsi signals on a 4 layer PCB.